style: 목표 상세 스타일, 파일위치 변경
